# Conflicts:
#	example/src/main/java/com/mercadopago/JsonSetupActivity.java
#	example/src/main/java/com/mercadopago/examples/checkout/CheckoutExampleActivity.java
#	sdk/src/main/java/com/mercadopago/CheckoutActivity.java
#	sdk/src/main/java/com/mercadopago/model/FeeDetail.java
#	sdk/src/main/java/com/mercadopago/model/PaymentMethod.java

+++ sdk/src/main/java/com/mercadopago/util/BitmapUtils.java
@@ -11,10 +11,10 @@ public class BitmapUtils {
     public static Bitmap scaleDown(Bitmap realImage, float maxImageSize,
                                    boolean filter) {
         float ratio = Math.min(
-                (float) maxImageSize / realImage.getWidth(),
-                (float) maxImageSize / realImage.getHeight());
-        int width = Math.round((float) ratio * realImage.getWidth());
-        int height = Math.round((float) ratio * realImage.getHeight());
+                maxImageSize / realImage.getWidth(),
+                maxImageSize / realImage.getHeight());
+        int width = Math.round(ratio * realImage.getWidth());
+        int height = Math.round(ratio * realImage.getHeight());
 
         Bitmap newBitmap = Bitmap.createScaledBitmap(realImage, width,
                 height, filter);
